delete jar only after full claim and withdraw, some test fixes
