+++ contract/src/jar/tests.rs
@@ -328,7 +328,8 @@ mod signature_tests {
 
         let alice_jars = context.contract.get_jars_for_account(alice);
 
-        assert_eq!(1, alice_jars.len());
+        assert_eq!(2, alice_jars.len());
+        assert_eq!(0, alice_jars.iter().find(|item| item.id.0 == 0).unwrap().principal.0);
         assert_eq!(
             1_000_000,
             alice_jars.iter().find(|item| item.id.0 == 1).unwrap().principal.0
