+++ contract/src/withdraw/mod.rs
@@ -1,2 +1,3 @@
 pub mod api;
+mod tests;
 pub mod view;
